create script for comm plot
